Move configuration to separate namespace
We want to have the ability to use environment variables, defaults, etc.
for configuration of dev-only components (e.g. figwheel or scss builds).
So, configuration goes in its own namespace, allowing us to define
a config for Figwheel in that namespace, but only actually use it to
bring Figwheel up in `dev/user.clj`.